**Refactor and Enhance Array Processing Features**
- **Added Sorting Functions:**
  - `sortByNameAscending(arrayToSort)`: Sorts an array of objects in ascending order based on the `name` property.
  - `sortByNameDescending(arrayToSort)`: Sorts an array of objects in descending order based on the `name` property.
  - Centralized sorting logic into a reusable helper function `sortByHelper` to reduce code duplication.

- **Implemented Filtering Functions:**
  - `filterByReviewPending(items)`: Filters an array of items to include only those with the `isReviewed` property set to "Pending review".
  - `filterByNotReviewed(items)`: Filters an array of items to include only those with the `isReviewed` property set to "Not reviewed".
  - Centralized filtering logic into a reusable helper function `filterByHelper` for consistent validation and filtering.

- **Enhanced Error Handling:**
  - Ensured all objects in the arrays have the required `name` or `isReviewed` property.
  - Added type checks and comprehensive error messages for better validation and debugging.

- **Refactored Code Structure:**
  - Moved `createProductTable` from `review.js` to the `component` folder.
  - Modified `createProductTable` to accept an order/item list that is processed by the new sorting and filtering functions before rendering.

- **Future Work still left:**
  - Implement sorting by oldest and newest items functions.
